Refactor nb-tester
